Fix stylesheet and JavaScript paths
On my side there was an error loading the style sheets and scripts. I changed the file paths so that it works on my side. Not sure if anyone was having the same problem. I also added "ui button" to the class of the search button to give it a more refined look.